py/groq-agent/requirements.txt
fastapi
uvicorn
langchain-groq
python-dotenv